+++ test/addons-napi/8_passing_wrapped/binding.cc
@@ -1,3 +1,4 @@
+#include <node_api.h>
 #include "myobject.h"
 #include "../common.h"
